Move hex functions out of test mod
